A simple Go program that just execs another program
A simple Go program that takes as an argument aother program and simply
execs it with the given arguments. In addition, there is the option of
applying secomp filters (in a strict or non-strict mode) based on an array
of strings that contains all the allowed system calls.

The strict mode will trap all system calls that have not been listed in
the array of strings, while the non-strict mode will just log them.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>